requirements.txt
asgiref==3.7.2
attrs==23.1.0
axios==0.4.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.4
django-allauth==0.57.0
django-cors-headers==4.2.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
outcome==1.3.0.post0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
